# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CIBOL · 实验室系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "lab"
        ],
        "summary": "Update Config",
        "description": "保存当前学期与组会默认（管理员）。无则建、有则更新。",
        "operationId": "update_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/announcements": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "List Announcements",
        "description": "当前生效的公告（未过期），按 pinned → level → 时间排序。\n管理员可见全部公告；普通成员只见 audience='all' 或 'students' 的公告。",
        "operationId": "list_announcements_api_announcements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementOut"
                  },
                  "type": "array",
                  "title": "Response List Announcements Api Announcements Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "lab"
        ],
        "summary": "Publish Announcement",
        "operationId": "publish_announcement_api_announcements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/announcements/{ann_id}/toggle-pin": {
      "post": {
        "tags": [
          "lab"
        ],
        "summary": "Toggle Pin",
        "operationId": "toggle_pin_api_announcements__ann_id__toggle_pin_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ann_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ann Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/announcements/{ann_id}": {
      "delete": {
        "tags": [
          "lab"
        ],
        "summary": "Remove Announcement",
        "operationId": "remove_announcement_api_announcements__ann_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ann_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ann Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "List Meetings",
        "description": "整学期组会排期（按日期升序），含报告人与在线会议信息。",
        "operationId": "list_meetings_api_meetings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeetingOut"
                  },
                  "type": "array",
                  "title": "Response List Meetings Api Meetings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/meetings/schedule": {
      "put": {
        "tags": [
          "lab"
        ],
        "summary": "Replace Schedule",
        "description": "整体保存组会排期（管理员）。\n\n按日期 upsert：表里有该日期则更新、没有则新增；表里多出的日期删除。\n匹配到日期的会议**保留其在线会议信息**（online_*），避免重排丢失已预约的腾讯会议。",
        "operationId": "replace_schedule_api_meetings_schedule_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeetingOut"
                  },
                  "type": "array",
                  "title": "Response Replace Schedule Api Meetings Schedule Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/feedback": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List Feedback",
        "description": "管理员查看全部匿名意见，按时间倒序。",
        "operationId": "list_feedback_api_feedback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedbackOut"
                  },
                  "type": "array",
                  "title": "Response List Feedback Api Feedback Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Feedback",
        "description": "任意登录成员匿名提交；不落任何身份信息。",
        "operationId": "submit_feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/feedback/unread-count": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Unread Count",
        "operationId": "unread_count_api_feedback_unread_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Unread Count Api Feedback Unread Count Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/feedback/{fb_id}/read": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Mark Feedback Read",
        "operationId": "mark_feedback_read_api_feedback__fb_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fb Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feedback/{fb_id}": {
      "delete": {
        "tags": [
          "feedback"
        ],
        "summary": "Delete Feedback",
        "description": "管理员删除意见（清理无效/违规条目）。",
        "operationId": "delete_feedback_api_feedback__fb_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fb Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnnLevel": {
        "type": "string",
        "enum": [
          "info",
          "important",
          "urgent"
        ],
        "title": "AnnLevel"
      },
      "AnnouncementCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 10000,
            "title": "Body",
            "default": ""
          },
          "level": {
            "$ref": "#/components/schemas/AnnLevel",
            "default": "info"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "audience": {
            "type": "string",
            "enum": [
              "all",
              "students"
            ],
            "title": "Audience",
            "default": "all"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AnnouncementCreate"
      },
      "AnnouncementOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "level": {
            "$ref": "#/components/schemas/AnnLevel"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Publishedat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "body",
          "level",
          "pinned",
          "audience",
          "author",
          "publishedAt"
        ],
        "title": "AnnouncementOut"
      },
...
      "ConfigIn": {
        "properties": {
          "semester": {
            "$ref": "#/components/schemas/SemesterIn"
          },
          "meetingDefault": {
            "$ref": "#/components/schemas/MeetingDefaultIn"
          }
        },
        "type": "object",
        "required": [
          "semester",
          "meetingDefault"
        ],
        "title": "ConfigIn"
      },
      "ConfigOut": {
        "properties": {
          "semester": {
            "$ref": "#/components/schemas/SemesterOut"
          },
          "meetingDefault": {
            "$ref": "#/components/schemas/MeetingDefaultOut"
          }
        },
        "type": "object",
        "required": [
          "semester",
          "meetingDefault"
        ],
        "title": "ConfigOut"
      },
...
      "FeedbackCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "FeedbackCreate"
      },
      "FeedbackOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body",
          "created_at",
          "read"
        ],
        "title": "FeedbackOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingDefaultIn": {
        "properties": {
          "weekday": {
            "type": "string",
            "title": "Weekday",
            "default": "周日"
          },
          "time": {
            "type": "string",
            "title": "Time",
            "default": ""
          },
          "place": {
            "type": "string",
            "title": "Place",
            "default": ""
          }
        },
        "type": "object",
        "title": "MeetingDefaultIn"
      },
      "MeetingDefaultOut": {
        "properties": {
          "weekday": {
            "type": "string",
            "title": "Weekday"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "place": {
            "type": "string",
            "title": "Place"
          }
        },
        "type": "object",
        "required": [
          "weekday",
          "time",
          "place"
        ],
        "title": "MeetingDefaultOut"
      },
...
      "MeetingOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          },
          "mo": {
            "type": "integer",
            "title": "Mo"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "mdLabel": {
            "type": "string",
            "title": "Mdlabel"
          },
          "dateLabel": {
            "type": "string",
            "title": "Datelabel"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "tone": {
            "type": "string",
            "title": "Tone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "正式报告"
          },
          "scored": {
            "type": "boolean",
            "title": "Scored",
            "default": true
          },
          "host": {
            "type": "string",
            "title": "Host",
            "default": ""
          },
          "time": {
            "type": "string",
            "title": "Time",
            "default": ""
          },
          "place": {
            "type": "string",
            "title": "Place",
            "default": ""
          },
          "online": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OnlineMeetingOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "presenters": {
            "items": {
              "$ref": "#/components/schemas/PresenterOut"
            },
            "type": "array",
            "title": "Presenters"
          }
        },
        "type": "object",
        "required": [
          "id",
          "y",
          "mo",
          "day",
          "mdLabel",
          "dateLabel",
          "type",
          "tone",
          "status",
          "presenters"
        ],
        "title": "MeetingOut"
      },
...
      "OnlineMeetingOut": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "OnlineMeetingOut"
      },
...
      "ScheduleIn": {
        "properties": {
          "meetings": {
            "items": {
              "$ref": "#/components/schemas/MeetingIn"
            },
            "type": "array",
            "title": "Meetings"
          },
          "scope_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope From"
          },
          "scope_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope To"
          }
        },
        "type": "object",
        "required": [
          "meetings"
        ],
        "title": "ScheduleIn"
      },
      "SemesterIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "short": {
            "type": "string",
            "title": "Short",
            "default": ""
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start",
          "end"
        ],
        "title": "SemesterIn"
      },
      "SemesterOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "short": {
            "type": "string",
            "title": "Short"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "name",
          "short",
          "start",
          "end"
        ],
        "title": "SemesterOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}